+++ HISTORY.md
@@ -26,6 +26,8 @@ Unpublished
 * The library name in the log context object is now `@expo/rudder-sdk-node`
 * The user agent string is now `expo-rudder-sdk-node/<version>`
 * Events are now removed from the event queue prior to sending a network request, rather than after
+* flush now guards against requests which are too large via configurable `maxFlushSizeInBytes` property
+* enqueue now limits the max events in the queue via configurable `maxQueueLength` property
 
 v1.0.8
 ==========================